refactor(Switch): remove the interface to index file
