--- src/components/inputs/Switch/interfaces/Switch.interface.ts
@@ -1,15 +1,0 @@
-import { Size } from "../types/Switch.Size.type";
-
-export interface ISwitchProps {
-  isDisabled?: boolean;
-  id: string;
-  name?: string;
-  value?: string;
-  size?: Size;
-  checked?: boolean;
-  handleChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  label: string;
-  margin: string;
-  padding: string;
-  disabled?: boolean;
-}
